% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcounts.R
\name{cmcounts}
\alias{cmcounts}
\title{Confusion Matrix Counts}
\usage{
cmcounts(data)
}
\arguments{
\item{data=dataframe}{with two binary variables}
}
\value{
counts of 00,01,10,11
}
\description{
Counts the frequencies for confusion matrix cells: 00,01,10,11.
}
\details{
Load a dataframe with two binary variables:  column 1 = "actual" condition and column 2 = "predicted" condition
}
\examples{

cmcounts(data=barretts)

}
\author{
Barbara Breeden
}
