% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.R
\name{conf}
\alias{conf}
\title{Diagnostic odds ratio 95\% confidence interval}
\usage{
conf(cm)
}
\arguments{
\item{cm=matrix}{:confusion matrix object from conmat() function.}
}
\value{
Diagnostic odds ratio 95\% confidence interval
}
\description{
Computes the 95\% confidence interval for diagnostic odds ratio.
}
\examples{

counts <- cmcounts(barretts)
cm <- conmat(counts)
conf <- conf(cm)
conf

}
\author{
Barbara Breeden
}
