% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conmat.R
\name{conmat}
\alias{conmat}
\title{Confusion Matrix}
\usage{
conmat(counts)
}
\arguments{
\item{counts=table}{(cmcounts() table) or vector containing frequencies of actual vs. predicted binary variables :  c(00,01,10,11)}
}
\value{
Labelled confusion matrix
}
\description{
Creates a confusion matrix.
}
\details{
Uses the table generated from cmcounts() function, or a vector of binary outcomes to create a labelled confusion matrix
}
\examples{

counts <- c(43,20,5,31) #frequencies of binary outcomes: 00,01,10,11
cm <- conmat(counts)
cm

counts <- cmcounts(data=barretts)
cm <- conmat(counts)
cm

}
\author{
Barbara Breeden
}
