% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Confusion Matrix Metrics}
\usage{
metrics(cm)
}
\arguments{
\item{cm=matrix}{:confusion matrix object from conmat() function.}
}
\value{
dataframe of metrics variables: P,N,PP,PN,TP,TN,FP,FN,n
}
\description{
Computes the metrics for the confusion matrix
}
\details{
(P)-Total positive, (N)-Total negative, (PP)-Predicted positive, (PN)-Predicted negative,
(TP)-True positive, (TN)-True negative, (FP)- False positive, (FN)-False negative, (n)-sample size
}
\examples{

counts <- cmcounts(barretts)
cm <- conmat(counts)
met <- metrics(cm)
met

}
\author{
Barbara Breeden
}
