% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdm.R
\name{pdm}
\alias{pdm}
\title{Prevalence Dependent Metrics}
\usage{
pdm(cm)
}
\arguments{
\item{cm=matrix}{:confusion matrix object from conmat() function.}
}
\value{
dataframe of metrics variables: Prevalence, PPV, NPV, F1, FDR, FOR, MCC
}
\description{
Computes the prevalence dependent metrics from the confusion matrix
}
\details{
Prevalence, (PPV)- positive predictive value, (NPV)- negative predictive value, (F1)- F1 score, (FDR)- false discovery rate,
(FOR)- false omission rate, (MCC), Matthews correlation coefficient
}
\examples{

counts <- cmcounts(barretts)
cm <- conmat(counts)
pdm <- pdm(cm)
pdm

}
\author{
Barbara Breeden
}
