% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pim.R
\name{pim}
\alias{pim}
\title{Prevalence Independent Metrics}
\usage{
pim(cm)
}
\arguments{
\item{cm=matrix}{:confusion matrix object from conmat() function.}
}
\value{
dataframe of metrics variables: ACC, TPR, TNR, BA, BM, FPR, FNR, PLR, NLR, DOR
}
\description{
Computes the prevalence independent metrics from the confusion matrix
}
\details{
(ACC)-Accuracy, (TPR)- True Positive rate, "Sensitivity", (TNR)- True Negative Rate, "Specificity", (BA)- Balanced Accuracy,
(BM)- Bookmaker informedness, (FPR)- False Positive Rate, (FNR)- False Negative Rate, (PLR)- Positive Likelihood Ratio,
(NLR)- Negative Likelihood Ratio, (DOR)- Diagnostic Odds Ratio
}
\examples{

counts <- cmcounts(barretts)
cm <- conmat(counts)
pim <- pim(cm)
pim

}
\author{
Barbara Breeden
}
